fixed the location
(cherry picked from commit 6605950983121a6348850994d3f131a1a30f2fc8)
