+++ backend/apps/owasp/management/commands/owasp_crp_recalculate_scores.py
@@ -2,7 +2,7 @@
 
 from django.core.management.base import BaseCommand, CommandError
 
-from apps.owasp.score_calculator import ContributionScoreCalculator
+from apps.owasp.models.crp.score_calculator import ContributionScoreCalculator
 
 
 class Command(BaseCommand):
