+++ backend/apps/owasp/models/crp/certificate.py
@@ -2,15 +2,20 @@
 
 from __future__ import annotations
 
+import logging
 import secrets
 
-from django.db import models
+from django.db import models, transaction
 from django.db.models import Q
 
 from apps.common.models import TimestampedModel
 from apps.github.models.user import User
+from apps.owasp.exceptions import CertificateIssuanceError
+from apps.owasp.models.crp.certificate_provider import CertificateProviderFactory
 from apps.owasp.models.crp.recognition_enums import TierChoices
 
+
+logger = logging.getLogger(__name__)
 
 CERTIFICATE_ID_ALPHABET = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ"
 CERTIFICATE_ID_LENGTH = 12
@@ -87,3 +92,53 @@ class Certificate(TimestampedModel):
         """Return human-readable representation."""
         status = "Revoked" if self.is_revoked else "Active"
         return f"{self.github_user.login} - {self.tier.upper()} Certificate ({status})"
+
+    @classmethod
+    @transaction.atomic
+    def issue_certificate(cls, user: User, score: int, tier: TierChoices) -> None:
+        """Issue a certificate for the user's current tier if one does not already exist.
+
+        Locks the User row to serialize concurrent issuances, checks for an existing
+        active certificate at this tier, and delegates to the configured provider.
+
+        Args:
+            user (User): The user to issue a certificate for.
+            score (int): The current contribution score of the user.
+            tier (TierChoices): The tier the user qualifies for.
+
+        Raises:
+            CertificateIssuanceError: If provider resolution or issuance fails.
+
+        """
+        # Lock the User row to serialize concurrent certificate issuances for this user
+        user = User.objects.select_for_update().get(id=user.id)
+
+        # Check if user already has an active certificate for this specific tier
+        if cls.objects.filter(
+            github_user=user,
+            tier=tier,
+            is_revoked=False,
+        ).exists():
+            return
+
+        try:
+            provider = CertificateProviderFactory.get_provider()
+        except ValueError as e:
+            logger.exception("Failed to resolve certificate provider")
+            raise CertificateIssuanceError from e
+
+        logger.info(
+            "Issuing %s certificate to user %s with score %s",
+            tier,
+            user.login,
+            score,
+        )
+        try:
+            provider.issue_certificate(user, score, tier)
+        except Exception as e:
+            logger.exception(
+                "Failed to issue %s certificate for user %s",
+                tier,
+                user.login,
+            )
+            raise CertificateIssuanceError from e
