--- backend/apps/owasp/services/__init__.py
@@ -1,1 +1,0 @@
-"""Service layer for Contributor Recognition Program."""
